% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_exploration.R
\name{explore_features}
\alias{explore_features}
\title{Plots repertoire features for binding versus non binding cells. This can be done both at the clonotype level or at the single cell level.}
\usage{
explore_features(features, per.sample)
}
\arguments{
\item{features}{List of dataframes containing the extracted features. This is the output of load_data function.}

\item{per.sample}{Logical indicating if the plots produced are separated by sample or aggregated.}
}
\value{
This function plots the mean mutation for each cell (divided in Heavy chain and Light chain mutations), gene usage, CDR3 length, and the ratio between coding versus non-coding mutations.
}
\description{
Plots repertoire features for binding versus non binding cells. This can be done both at the clonotype level or at the single cell level.
}
\examples{
\dontrun{
check_explore_features <- explore_features(features = output.load_data)
}

}
