% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App.R
\name{label_interactive}
\alias{label_interactive}
\title{Launches an interactive web application for labeling cells and clonotypes based on the results of an ELISA experiment, and adds the possibility to add an affintiy value for affinity prediction.}
\usage{
label_interactive(features)
}
\arguments{
\item{features}{List of dataframes containing the extracted features. This is the output of load_data function.}
}
\description{
Launches an interactive web application for labeling cells and clonotypes based on the results of an ELISA experiment, and adds the possibility to add an affintiy value for affinity prediction.
}
\examples{
\dontrun{
label_interactive(features = output.load_data)
}

}
