% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Analyzes and processes the repertoire sequencing data from cellranger vdj. This provides information on the single-cell level for each clone.}
\usage{
load_data(
  VDJ.directory,
  clonotype.level = FALSE,
  filter.prod = TRUE,
  filter.confidence = TRUE,
  filter.is.cell = TRUE,
  red.method = "pca",
  n.dims = 1:10
)
}
\arguments{
\item{VDJ.directory}{Character vector with each element containing the path to the output of cellranger vdj runs. To integrate the repertoire with transcriptome data GEX data can by provided inside the corresponding folder. The name of the GEX folder must contain "gex" in his name (case insensitive). This pipeline assumes that the output file names have not been changed from the default 10x settings in the /outs/ folder.}

\item{red.method}{Method of dimensionality reduction used. Options are "pca", "tsne", "umap".}

\item{n.dims}{Number of dimension to be used in the chosen reduction. Default is 20.}

\item{filtered.prod}{Logical indicating if the contigs file should be filtered by productive cells. Default set to TRUE}

\item{filtered.confidence}{Logical indicating if the contigs file should be filtered by contigs with high_confidence. Default set to TRUE}

\item{filtered.is.cell}{Logical indicating if the contigs file should be filtered by sequences that only come from cells. Default set to TRUE}
}
\value{
Returns a list of dataframes containing information on the cell or clonotype level.
}
\description{
Analyzes and processes the repertoire sequencing data from cellranger vdj. This provides information on the single-cell level for each clone.
}
\examples{
\dontrun{
check_load_data <- load_data(VDJ.directory = "path/to/cellranger/outs/", clonotype.level = FALSE)
}

}
